Moved stack operations to own module
